fix(daisi): Adapt data types and casting

+++ daisi/src/cpps/common/scenariofile/amr_description_scenario.h
@@ -44,7 +44,7 @@ struct AmrLoadHandlingScenario {
   double load_time;
   double unload_time;
   std::string load_carrier;
-  double max_payload;
+  float max_payload;
 
   void parse(const YAML::Node &node) {
     SERIALIZE_VAR(load_time);
